+++ src/data_server.hpp
@@ -30,24 +30,8 @@
 #include "sql.hpp"
 
 class data_server {
-    // static data is loaded once. They are known full sets like SchoolUnit
-    // All are loaded initially.
-    // Also generated object like Employment and Activity are static.
-    std::map<std::string, std::shared_ptr<object_list>> static_data;
-    // dynamic data is objects like User, they are loaded as a consequence of loading
-    // groups and generated data.
-    // The distinction is if we can say "it is loaded or not" v.s. it grows as we load
-    // other data
-    std::map<std::string, std::shared_ptr<object_list>> dynamic_data;
-
-    // when relational objects are loaded they will later be requested by
-    // another key than GUID.
-    // This requires a search of all objects by attribute. This cache intercepts that and
-    // is indexed by <type><attribute><id>. It holds weak pointers
-    std::map<std::string, std::weak_ptr<base_object>> alt_key_cache;
-
-    string_vector static_types;
-    string_vector dynamic_types;
+    // Loaded data, arranged by type
+    std::map<std::string, std::shared_ptr<object_list>> data;
 
     std::unique_ptr<ldap_wrapper> ldap;
     std::unique_ptr<csv_store> csv;
@@ -56,14 +40,6 @@ class data_server {
 
     data_server() = default;
 
-    bool isStatic(const std::string &type) const {
-        for (auto && v : static_types) {
-            if (v == type)
-                return true;
-        }
-        return false;
-    }
-
 public:
     static data_server &instance() {
         static data_server s;
@@ -71,20 +47,15 @@ public:
     }
 
     void clear() {
-        static_types.clear();
-        dynamic_types.clear();
-        static_data.clear();
-        dynamic_data.clear();
+        data.clear();
         ldap.reset();
         csv.reset();
     }
 
     bool empty() {
-        return static_data.empty() && dynamic_data.empty();
+        return data.empty();
     }
 
-//	std::shared_ptr<base_object>
-//	get_object_by_attribute(const std::string &type, const std::string &attrib, const std::string &value);
     std::shared_ptr<base_object>
     find_object_by_attribute(const std::string &type, const std::string &attrib, const std::string &value);
 
@@ -92,19 +63,9 @@ public:
 
     std::shared_ptr<object_list> get_by_type(const std::string &type) const;
 
-    /**
-     * dependent data that is loaded in its entirety once
-     * @param type
-     * @return
-     */
-    std::shared_ptr<object_list> get_static_by_type(const std::string &type);
-
-
     bool load(std::shared_ptr<sql::plugin> sql_plugin);
 
     void preload();
-
-    void cache_relation(const std::string& key, std::weak_ptr<base_object> object);
 
     void add(const std::string& type, std::shared_ptr<object_list> list);
 
@@ -125,19 +86,7 @@ public:
     }
 
 private:
-    /**
-     * static_data is loaded once
-     * @param type
-     * @param list
-     */
-    void add_static(const std::string &type, std::shared_ptr<object_list> list);
-
-    /**
-     * add dynamic data, this is appended to as static data is loaded
-     * @param type
-     * @param list
-     */
-    void add_dynamic(const std::string &type, std::shared_ptr<object_list> list);
+    void add_internal(const std::string &type, std::shared_ptr<object_list> list);
 
     void filter_orphans();